+++ lib/demo.dart
@@ -1,83 +1,124 @@
-import 'main.dart';
 import 'package:flutter/material.dart';
+import 'chat.dart';
+import 'main.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
-class Demo extends StatelessWidget {
-  const Demo({Key? key}) : super(key: key);
+class ChatDetailPage extends StatefulWidget{
+  @override
+  _ChatDetailPageState createState() => _ChatDetailPageState();
+}
+
+class _ChatDetailPageState extends State<ChatDetailPage> {
+
+  final CollectionReference _protips =
+  FirebaseFirestore.instance.collection('protips');
 
   @override
   Widget build(BuildContext context) {
-    return const MaterialApp(
-      debugShowMaterialGrid: false,
-      debugShowCheckedModeBanner: false,
-      home: DemoPage(),
-    );
-  }
-}
-
-class DemoPage extends StatefulWidget {
-  const DemoPage({Key? key}) : super(key: key);
-
-  @override
-  _DemoState createState() => _DemoState();
-}
-
-class _DemoState extends State<DemoPage> {
-  final CollectionReference _achievements =
-  FirebaseFirestore.instance.collection('achievements');
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      // Remove the debug banner
-      debugShowCheckedModeBanner: false,
-      debugShowMaterialGrid: false,
-      theme: ThemeData(
-        primarySwatch: Colors.red,
-      ),
-      home: Scaffold(
+    return Scaffold(
         appBar: AppBar(
-          leading: IconButton(
-            icon: Icon(Icons.arrow_back_ios, color: Colors.black,),
-            onPressed: ()=>Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (_)=>MyHomePage(title: 'Tween'))),
+          elevation: 0,
+          automaticallyImplyLeading: false,
+          backgroundColor: Colors.white,
+          flexibleSpace: SafeArea(
+            child: Container(
+              padding: EdgeInsets.only(right: 16),
+              child: Row(
+                children: <Widget>[
+                  IconButton(
+                    onPressed: ()=>Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (_)=>MyHomePage(title: 'Tween'))),
+                    icon: Icon(Icons.arrow_back,color: Colors.black,),
+                  ),
+                  SizedBox(width: 2,),
+                  const CircleAvatar(
+                    backgroundImage: NetworkImage("https://assets4.lottiefiles.com/packages/lf20_ik4jyixs.json"),
+                    maxRadius: 20,
+                  ),
+                  SizedBox(width: 12,),
+                  Expanded(
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        Text("Community",style: TextStyle( fontSize: 16 ,fontWeight: FontWeight.w600),),
+                      ],
+                    ),
+                  ),
+                ],
+              ),
+            ),
           ),
-          title: const Text('Demo'),
-          centerTitle: true,
         ),
-        // Using StreamBuilder to display all products from Firestore in real-time
-        body: StreamBuilder(
-          stream: _achievements.snapshots(),
-          builder: (context, AsyncSnapshot<QuerySnapshot> streamSnapshot) {
-            if (streamSnapshot.hasData) {
-              return GridView.count(
-                shrinkWrap: true,
-                physics: ClampingScrollPhysics(),
-                crossAxisCount: 1,
-                childAspectRatio: 3.5,
-                children: streamSnapshot.data!.docs.map((document) {
-                  return GestureDetector(
-                    child: Card(
-                      elevation: 3.0,
-                      margin: const EdgeInsets.all(10),
-                      child: Row(
-                        children: [
-                          Text(document['title'], style: TextStyle(color: Colors.black54,),),
-                          Text(document['desc'], style: TextStyle(color: Colors.black54,),),
-                          Text(document['condition'], style: TextStyle(color: Colors.black54,),),
-                        ],
+      body: Stack(
+        children: <Widget>[
+          ListView.builder(
+            itemCount: messages.length,
+            shrinkWrap: true,
+            padding: EdgeInsets.only(top: 10,bottom: 10),
+            physics: NeverScrollableScrollPhysics(),
+            itemBuilder: (context, index){
+              return Container(
+                padding: EdgeInsets.only(left: 14,right: 14,top: 10,bottom: 10),
+                child: Align(
+                  alignment: (messages[index].messageType == "receiver"?Alignment.topLeft:Alignment.topRight),
+                  child: Container(
+                    decoration: BoxDecoration(
+                      borderRadius: BorderRadius.circular(20),
+                      color: (messages[index].messageType  == "receiver"?Colors.grey.shade200:Colors.blue[200]),
+                    ),
+                    padding: EdgeInsets.all(16),
+                    child: Text(messages[index].messageContent, style: TextStyle(fontSize: 15),),
+                  ),
+                ),
+              );
+            },
+          ),
+          Align(
+            alignment: Alignment.bottomLeft,
+            child: Container(
+              padding: EdgeInsets.only(left: 10,bottom: 10,top: 10),
+              height: 60,
+              width: double.infinity,
+              color: Colors.white,
+              child: Row(
+                children: <Widget>[
+                  GestureDetector(
+                    onTap: (){
+                    },
+                    child: Container(
+                      height: 30,
+                      width: 30,
+                      decoration: BoxDecoration(
+                        color: Colors.lightBlue,
+                        borderRadius: BorderRadius.circular(30),
+                      ),
+                      child: Icon(Icons.add, color: Colors.white, size: 20, ),
+                    ),
+                  ),
+                  SizedBox(width: 15,),
+                  Expanded(
+                    child: TextField(
+                      decoration: InputDecoration(
+                          hintText: "Write message...",
+                          hintStyle: TextStyle(color: Colors.black54),
+                          border: InputBorder.none
                       ),
                     ),
-                  );
-                }).toList(),
-              );
-            }
+                  ),
+                  SizedBox(width: 15,),
+                  FloatingActionButton(
+                    onPressed: (){},
+                    child: Icon(Icons.send,color: Colors.white,size: 18,),
+                    backgroundColor: Colors.blue,
+                    elevation: 0,
+                  ),
+                ],
 
-            return const Center(
-              child: CircularProgressIndicator(),
-            );
-          },
-        ),
+              ),
+            ),
+          ),
+        ],
       ),
     );
   }
