+++ lib/premium.dart
@@ -1,9 +1,41 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
 import 'main.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:cloud_firestore/cloud_firestore.dart';
 
 class Premium extends StatelessWidget {
   const Premium({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      debugShowMaterialGrid: false,
+      debugShowCheckedModeBanner: false,
+      home: ProPage(),
+    );
+  }
+}
+
+class ProPage extends StatefulWidget {
+  const ProPage({Key? key}) : super(key: key);
+
+  @override
+  _ProState createState() => _ProState();
+}
+
+class _ProState extends State<ProPage> {
+  //CollectionReference
+  final _users = FirebaseFirestore.instance.collection('user');
+
+  // Deleteing a product by id
+  Future<void> ToPro() async {
+    FirebaseFirestore.instance.collection('user').doc("tS2qOJ5Pe8WY1sd6EvaQ7WdnYYP2").update({'isPro': true});
+
+    // Show a snackbar
+    ScaffoldMessenger.of(context).showSnackBar(const SnackBar(
+        content: Text('You have successfully became a PRO user.')));
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -24,7 +56,9 @@ class Premium extends StatelessWidget {
           elevation: 0.0,
           leading: IconButton(
             icon: Icon(Icons.arrow_back_ios, color: Colors.black,),
-            onPressed: ()=>Navigator.of(context).pushReplacement(MaterialPageRoute(builder: (_)=>MyHomePage(title: 'Tween'))),
+            onPressed: () =>
+                Navigator.of(context).pushReplacement(MaterialPageRoute(
+                    builder: (_) => MyHomePage(title: 'Tween'))),
           ),
         ),
         body: SingleChildScrollView(
@@ -56,7 +90,10 @@ class Premium extends StatelessWidget {
                         child: Row(
                           mainAxisAlignment: MainAxisAlignment.center,
                           children: [
-                            Text("Premium", style:TextStyle(color: Colors.black, fontWeight: FontWeight.w500, fontSize: 16.0),),
+                            Text("Premium", style: TextStyle(
+                                color: Colors.black,
+                                fontWeight: FontWeight.w500,
+                                fontSize: 16.0),),
                             //Text("Total Achievements Earned", style:TextStyle(color: Colors.white, fontWeight: FontWeight.w500, fontSize: 16.0),),
                             //Text("Global", style:TextStyle(color: Colors.white, fontWeight: FontWeight.w500, fontSize: 16.0),),
                           ],
@@ -83,7 +120,10 @@ class Premium extends StatelessWidget {
                   child: Padding(
                     padding: const EdgeInsets.all(2.0),
                     child: Container(
-                      height: MediaQuery.of(context).size.height * 0.75,
+                      height: MediaQuery
+                          .of(context)
+                          .size
+                          .height * 0.75,
                       decoration: BoxDecoration(
                         borderRadius: BorderRadius.only(
                           topLeft: Radius.circular(20.0),
@@ -92,10 +132,26 @@ class Premium extends StatelessWidget {
                         color: Colors.white,
                       ),
                       child: GridView.count(
-                        crossAxisCount: 3,
-                        childAspectRatio: 0.8,
+                        crossAxisCount: 1,
+                        childAspectRatio: 2,
                         children: [
-
+                          Card(
+                            elevation: 3.0,
+                            margin: const EdgeInsets.all(10),
+                            child: Column(
+                              mainAxisAlignment: MainAxisAlignment.center,
+                              crossAxisAlignment: CrossAxisAlignment.center,
+                              children: [
+                                Text("Benefits of being a PRO user", style: TextStyle(color: Colors.black54, fontWeight: FontWeight.bold,),),
+                                Text("No Ads", style: TextStyle(color: Colors.black54,),),
+                                Text("Unlock more achievements and the ability to create new achievements", style: TextStyle(color: Colors.black54,),),
+                                IconButton(
+                                    onPressed: () => ToPro(),
+                                    icon: Icon(Icons.add_shopping_cart),
+                                ),
+                              ],
+                            ),
+                          ),
                         ],
                       ),
                     ),
