+++ pubspec.yaml
@@ -36,9 +36,14 @@ dependencies:
   cupertino_icons: ^1.0.2
   numberpicker: ^2.1.1
   lottie: ^1.2.1
+  fluttertoast: ^8.0.9
   firebase_auth: ^3.3.20
   firebase_core: ^1.19.1
   cloud_firestore: ^3.1.18
+  #paypal
+  http_auth: ^1.0.1
+  webview_flutter: ^3.0.4
+  shared_preferences: ^2.0.15
 
 dev_dependencies:
   flutter_test:
